+++ CHANGELOG.md
@@ -6,5 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.0.26] - 2016-07-19
 ### Added
 - FEATURE: kumo now supports stack policies, see README for details
+- FEATURE: kumo now displays changes in CloudFormation template parameters
+- FIX: prettify output
 - FIX: removed debug output
 - FIX: removed some unneccessary import validations
+- FIX: kumo will now exit when importing a cloudformation.py not from your current working directory